Tidligere var det en sirkulær avhengighet, der restUtils.ts hentet logErrorMessage fra innsynsDataActions.ts og innsynsDataActions.ts hentete REST_STATUS fra restUtils.ts Dette gjorde blandt annet at tester man ønsker å lage med en import fra en av disse vil faile. Generelt kan sirkulære avhengigheter også få uventede konsekvenser.
Fikset det ved å dra loggActions.ts ut fra innsynsDataActions.ts slik at restUtils.ts ikke lenger har en avhengighet til innsynsDataActions.ts
