+++ src/redux/innsynsdata/innsynsDataActions.ts
@@ -1,5 +1,5 @@
 import {Dispatch} from "redux";
-import {fetchToJson, HttpStatus, REST_STATUS, fetchPost} from "../../utils/restUtils";
+import {fetchToJson, HttpStatus, REST_STATUS} from "../../utils/restUtils";
 import {
     InnsynsdataActionTypeKeys,
     InnsynsdataSti,
@@ -10,6 +10,7 @@ import {
     skalViseFeilside,
     oppdaterOppgaveState,
 } from "./innsynsdataReducer";
+import {logErrorMessage} from "./loggActions";
 
 export const innsynsdataUrl = (fiksDigisosId: string, sti: string): string => `/innsyn/${fiksDigisosId}/${sti}`;
 
@@ -101,39 +102,6 @@ export function hentSaksdetaljer(fiksDigisosId: string, visFeilSide?: boolean) {
             });
     };
 }
-
-const LOG_URL = "/info/logg";
-
-export function logInfoMessage(message: string, navCallId?: string) {
-    fetchPost(LOG_URL, JSON.stringify(createLogEntry(message, "INFO", navCallId)))
-        .then(() => {})
-        .catch(() => {
-            return; // Not important to handle those errors
-        });
-}
-
-export function logErrorMessage(message: string, navCallId?: string) {
-    fetchPost(LOG_URL, JSON.stringify(createLogEntry(message, "ERROR", navCallId)))
-        .then(() => {})
-        .catch(() => {
-            return; // Not important to handle those errors
-        });
-}
-
-function createLogEntry(message: string, level: LogLevel, navCallId?: string) {
-    return {
-        level: level,
-        message: message,
-        jsFileUrl: "",
-        lineNumber: "",
-        columnNumber: "",
-        url: window.location.href,
-        userAgent: window.navigator.userAgent,
-        loggenGjelderForCallId: navCallId,
-    };
-}
-
-type LogLevel = "ERROR" | "WARN" | "INFO";
 
 export const setOppgaveVedleggopplastingFeilet = (status: boolean) => ({
     type: InnsynsdataActionTypeKeys.OPPGAVE_VEDLEGSOPPLASTING_FEILET,
